requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
openai==1.35.0
chromadb==0.5.0
python-dotenv==1.0.1
jinja2==3.1.4
python-multipart==0.0.9
aiofiles==23.2.1
pydantic==2.7.4
